#2393 - Move partial index support to SQL Server 2008+
