+++ tests/Doctrine/Tests/DBAL/Platforms/SQLServer2008PlatformTest.php
@@ -19,4 +19,9 @@ class SQLServer2008PlatformTest extends AbstractSQLServerPlatformTestCase
                 array())
         );
     }
+
+    public function getGenerateUniqueIndexSql()
+    {
+        return 'CREATE UNIQUE INDEX index_name ON test (test, test2) WHERE test IS NOT NULL AND test2 IS NOT NULL';
+    }
 }
